+++ Project/holssi_BE/src/main/java/org/example/holssi_be/dto/AuthDTO.java
@@ -1,4 +1,4 @@
-package org.example.holssi_be.dto.auth;
+package org.example.holssi_be.dto;
 
 import jakarta.validation.constraints.NotEmpty;
 import lombok.Getter;
