--- Project/holssi_BE/src/main/java/org/example/holssi_be/dto/auth/IdentifierDTO.java
@@ -1,9 +1,0 @@
-package org.example.holssi_be.dto.auth;
-
-import lombok.Data;
-
-@Data
-public class IdentifierDTO {
-    private String primaryKey;  // email
-    private String identifier;
-}
